[fix] Fix RotaryEmbedding coding conflict when half-precision has integer part limitation.

+++ megatron/core/models/common/embeddings/rotary_pos_embedding.py
@@ -62,9 +62,9 @@ class RotaryEmbedding(nn.Module):
         """
         seq = torch.arange(max_seq_len, device=self.inv_freq.device) + offset
         if self.seq_len_interpolation_factor is not None:
-            seq = seq.type_as(self.inv_freq)
+            seq = seq.to(torch.float32)
             seq *= 1 / self.seq_len_interpolation_factor
-        freqs = einsum('i , j -> i j', seq.type_as(self.inv_freq), self.inv_freq)
+        freqs = einsum('i , j -> i j', seq.to(torch.float32), self.inv_freq)
         # first part even vector components, second part odd vector components,
         #  2 * dim in dimension size
         emb = torch.cat((freqs, freqs), dim=-1)
@@ -73,7 +73,7 @@ class RotaryEmbedding(nn.Module):
         if parallel_state.get_context_parallel_world_size() > 1:
             # slice rotary_pos_emb along sequence dimension and select the parition of the current CP rank
             emb = get_pos_emb_on_this_cp_rank(emb, 0)
-        return emb
+        return emb.type_as(self.inv_freq)
 
     def _load_from_state_dict(self, state_dict, prefix, *args, **kwargs):
         state_dict.pop(f'{prefix}inv_freq', None)
